feat: create usedargdrop custom hook
