+++ library/lib/App.tsx
@@ -1,4 +1,3 @@
-import { DragEvent } from "react"
 import {
   ReactFlow,
   ReactFlowProvider,
@@ -14,20 +13,16 @@ import {
   useNodesState,
   addEdge,
   useEdgesState,
-  useReactFlow,
-  type Node,
   Connection,
 } from "@xyflow/react"
-
 import "@xyflow/react/dist/style.css"
 import { MAX_SCALE_TO_ZOOM_IN, MIN_SCALE_TO_ZOOM_OUT } from "./contants"
 import { defaultEdges, defaultNodes } from "./initialElements"
 import "@/styles/app.css"
 import { Sidebar } from "@/components"
 import { useCallback } from "react"
-import { DropNodeData } from "@/types"
-import { generateUUID } from "@/utils"
 import { nodeTypes } from "./nodes"
+import { useDragDrop } from "./hooks"
 
 const defaultEdgeOptions: DefaultEdgeOptions = {
   type: "smoothstep",
@@ -40,48 +35,12 @@ interface AppProps {
 }
 
 function App({ onReactFlowInit }: AppProps) {
-  const [nodes, setNodes, onNodesChange] = useNodesState(defaultNodes)
+  const [nodes, , onNodesChange] = useNodesState(defaultNodes)
   const [edges, setEdges, onEdgesChange] = useEdgesState(defaultEdges)
-  const { screenToFlowPosition } = useReactFlow()
-
+  const { onDrop, onDragOver } = useDragDrop()
   const onConnect = useCallback(
     (params: Connection) => setEdges((eds) => addEdge(params, eds)),
     []
-  )
-
-  const onDragOver = useCallback((event: DragEvent) => {
-    event.preventDefault()
-    event.dataTransfer.dropEffect = "move"
-  }, [])
-
-  const onDrop = useCallback(
-    (event: DragEvent) => {
-      event.preventDefault()
-      const data = JSON.parse(
-        event.dataTransfer.getData("text/plain")
-      ) as DropNodeData
-
-      // check if the dropped element is valid
-      if (!data.type) {
-        return
-      }
-
-      const position = screenToFlowPosition({
-        x: event.clientX,
-        y: event.clientY,
-      })
-      const newNode: Node = {
-        width: 200,
-        height: 110,
-        id: generateUUID(),
-        type: data.type,
-        position,
-        data: data.data,
-      }
-
-      setNodes((nds) => nds.concat(newNode))
-    },
-    [screenToFlowPosition]
   )
 
   return (
@@ -95,8 +54,8 @@ function App({ onReactFlowInit }: AppProps) {
         onNodesChange={onNodesChange}
         onEdgesChange={onEdgesChange}
         onConnect={onConnect}
-        onDrop={onDrop}
         onDragOver={onDragOver}
+        onDrop={onDrop}
         connectionLineType={ConnectionLineType.SmoothStep}
         connectionMode={ConnectionMode.Loose}
         fitView
